Feat/maps activity (#49)
* add maps activity

* add tanggal mulai field

* change to range picker

* change order usign start and end date

* complete update image profil

* location selected but button still not worked

* add maps activity select location and validation

* reformat date start

* change base url

* adjut padding control gmaps

+++ app/src/main/java/com/example/helperstartup/Model/service/ResponseApi/PostTransactionResponse.kt
@@ -31,7 +31,7 @@ data class DataTransaction(
 	val address: String,
 
 	@field:SerializedName("lng")
-	val lng: Int,
+	val lng: Double,
 
 	@field:SerializedName("upload")
 	val upload: String,
@@ -67,7 +67,7 @@ data class DataTransaction(
 	val menuId: Int,
 
 	@field:SerializedName("lat")
-	val lat: Int,
+	val lat: Double,
 
 	@field:SerializedName("status")
 	val status: String
